fix: harden feed URLs, stabilize modal selection, relocate tests and update badge
