+++ internal/ui/model.go
@@ -146,6 +146,9 @@ type Model struct {
 	// persists in settings so a dismissed message never comes back.
 	dismissed    map[string]bool
 	noticeCursor int
+	// whatsNewVersion mirrors the persisted whats_new_version setting so
+	// the notices list, rebuilt every frame, never reads the database.
+	whatsNewVersion string
 	// feedMessages is the remote message feed, refreshed on the update
 	// tick and folded into the notices next to the built-in ones.
 	feedMessages []feed.Message
@@ -411,6 +414,7 @@ func newModel(cfg config.Config, st *store.Store, driver *tmux.Driver, engine *s
 		mode:            modeList,
 		update:          updateInfo{version: version},
 		dismissed:       loadDismissed(st),
+		whatsNewVersion: loadWhatsNewVersion(st),
 	}
 }
 
@@ -837,13 +841,10 @@ func (m *Model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 		if msg.failed {
 			return m, nil
 		}
-		// A release found while the notices modal is open prepends a new
-		// entry; the cursor follows so the selected notice stays selected.
-		if m.mode == modeNotices && m.update.latest == "" && msg.latest != "" {
-			m.noticeCursor++
-		}
-		m.update.latest = msg.latest
-		m.update.url = msg.url
+		m.keepNoticeSelection(func() {
+			m.update.latest = msg.latest
+			m.update.url = msg.url
+		})
 		return m, nil
 
 	case updateTickMsg:
@@ -851,7 +852,7 @@ func (m *Model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 
 	case feedMsg:
 		if !msg.failed {
-			m.feedMessages = msg.messages
+			m.keepNoticeSelection(func() { m.feedMessages = msg.messages })
 		}
 		return m, nil
