+++ internal/ui/notices_test.go
@@ -7,6 +7,10 @@ import (
 
 	"github.com/YoanWai/agent-manager/internal/feed"
 	"github.com/YoanWai/agent-manager/internal/store"
+	"github.com/YoanWai/agent-manager/internal/sysstat"
+	tea "github.com/charmbracelet/bubbletea"
+	"github.com/charmbracelet/lipgloss"
+	"github.com/charmbracelet/x/ansi"
 )
 
 func noticeStore(t *testing.T) *store.Store {
@@ -21,9 +25,10 @@ func noticeStore(t *testing.T) *store.Store {
 
 func noticeModel(st *store.Store, version string) *Model {
 	return &Model{
-		store:     st,
-		update:    updateInfo{version: version},
-		dismissed: loadDismissed(st),
+		store:           st,
+		update:          updateInfo{version: version},
+		dismissed:       loadDismissed(st),
+		whatsNewVersion: loadWhatsNewVersion(st),
 	}
 }
 
@@ -204,4 +209,274 @@ func contains(list []string, want string) bool {
 		}
 	}
 	return false
+}
+
+func footModel(t *testing.T) *Model {
+	t.Helper()
+	m := noticeModel(noticeStore(t), "v0.2.0")
+	m.snap = sysstat.Snapshot{
+		CPUPercent: 42, CPUOK: true,
+		MemPercent: 63, MemOK: true, MemUsed: 10 << 30, MemTotal: 16 << 30,
+		DiskPercent: 71, DiskOK: true, DiskFree: 120 << 30,
+	}
+	return m
+}
+
+func TestRailFootPutsMessagesRightOfComputer(t *testing.T) {
+	m := footModel(t)
+	lines := m.railFootLines(70)
+
+	joined := ansi.Strip(strings.Join(lines, "\n"))
+	if !strings.Contains(joined, "messages") {
+		t.Fatalf("want a messages card, got %q", joined)
+	}
+	if !strings.Contains(joined, "welcome") {
+		t.Fatalf("want the welcome banner, got %q", joined)
+	}
+
+	for _, line := range lines {
+		clean := ansi.Strip(line)
+		if !strings.Contains(clean, "messages") {
+			continue
+		}
+		if strings.Index(clean, "messages") < strings.Index(ansi.Strip(lines[0]), "computer") {
+			t.Fatalf("messages must sit right of computer, got %q", clean)
+		}
+		return
+	}
+	t.Fatal("MESSAGES header row not found")
+}
+
+func TestRailFootNarrowDropsMessages(t *testing.T) {
+	m := footModel(t)
+	lines := m.railFootLines(34)
+	joined := ansi.Strip(strings.Join(lines, "\n"))
+	if strings.Contains(joined, "messages") {
+		t.Fatalf("narrow rail should keep only the meters, got %q", joined)
+	}
+	if !strings.Contains(joined, "computer") {
+		t.Fatalf("meters must survive, got %q", joined)
+	}
+}
+
+func TestRailFootAllDismissedShowsOnlyMeters(t *testing.T) {
+	m := footModel(t)
+	for _, n := range m.activeNotices() {
+		m.dismissNotice(n.id)
+	}
+	joined := ansi.Strip(strings.Join(m.railFootLines(70), "\n"))
+	if strings.Contains(joined, "messages") {
+		t.Fatalf("no notices means no panel, got %q", joined)
+	}
+}
+
+func TestRailFootCardBorderAndFit(t *testing.T) {
+	m := footModel(t)
+	lines := m.railFootLines(90)
+	joined := ansi.Strip(strings.Join(lines, "\n"))
+	for _, corner := range []string{"╭", "╮", "╰", "╯"} {
+		if !strings.Contains(joined, corner) {
+			t.Fatalf("card border missing %q:\n%s", corner, joined)
+		}
+	}
+	if !strings.Contains(strings.Join(lines, "\n"), bgSeq(noticeCardHex())) {
+		t.Fatal("card interior missing its fill")
+	}
+	for i, line := range lines {
+		if !strings.Contains(ansi.Strip(line), "│") {
+			t.Fatalf("row %d missing the separator: %q", i, ansi.Strip(line))
+		}
+	}
+
+	var top string
+	for _, line := range lines {
+		if strings.Contains(ansi.Strip(line), "╭") {
+			top = line
+		}
+	}
+	if got := lipgloss.Width(top); got >= 90 {
+		t.Fatalf("card must hug its content, top border spans %d of 90", got)
+	}
+}
+
+func TestRailFootLinesFitWidth(t *testing.T) {
+	m := footModel(t)
+	m.update.latest = "v0.9.9"
+	m.update.url = "https://example.com"
+	for _, width := range []int{40, 55, 70} {
+		for _, line := range m.railFootLines(width) {
+			if got := lipgloss.Width(line); got > width {
+				t.Fatalf("width %d: line overflows at %d: %q", width, got, ansi.Strip(line))
+			}
+		}
+	}
+}
+
+func key(s string) tea.KeyMsg {
+	switch s {
+	case "enter":
+		return tea.KeyMsg{Type: tea.KeyEnter}
+	case "esc":
+		return tea.KeyMsg{Type: tea.KeyEscape}
+	case "down":
+		return tea.KeyMsg{Type: tea.KeyDown}
+	case "up":
+		return tea.KeyMsg{Type: tea.KeyUp}
+	}
+	return tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune(s)}
+}
+
+func modalModel(t *testing.T) *Model {
+	t.Helper()
+	m := footModel(t)
+	m.width, m.height = 100, 34
+	m.mode = modeNotices
+	return m
+}
+
+func TestOpenNoticesFromList(t *testing.T) {
+	m := footModel(t)
+	m.width, m.height = 100, 34
+	m.mode = modeList
+	m.handleKey(key("M"))
+	if m.mode != modeNotices {
+		t.Fatalf("M should open the notices modal, mode=%v", m.mode)
+	}
+}
+
+func TestNoticesViewListsAndDetails(t *testing.T) {
+	m := modalModel(t)
+	frame := ansi.Strip(m.View())
+	for _, want := range []string{"messages", "Welcome to agent-manager", "Found a bug?", "dismiss"} {
+		if !strings.Contains(frame, want) {
+			t.Fatalf("modal missing %q:\n%s", want, frame)
+		}
+	}
+
+	var widths []int
+	for _, line := range strings.Split(frame, "\n") {
+		trimmed := strings.TrimRight(line, " ")
+		if strings.ContainsAny(trimmed, "╭╰│") {
+			widths = append(widths, len([]rune(trimmed)))
+		}
+	}
+	if len(widths) == 0 {
+		t.Fatal("no frame rows found")
+	}
+	for i, width := range widths {
+		if width != widths[0] {
+			t.Fatalf("frame rows must align: row %d is %d, first is %d\n%s", i, width, widths[0], frame)
+		}
+	}
+}
+
+func TestNoticesDismissAdvancesAndCloses(t *testing.T) {
+	m := modalModel(t)
+	total := len(m.activeNotices())
+	for i := 0; i < total; i++ {
+		m.handleNoticesKey(key("x"))
+	}
+	if m.mode != modeList {
+		t.Fatal("dismissing the last notice should close the modal")
+	}
+	if len(m.activeNotices()) != 0 {
+		t.Fatal("all notices should be gone")
+	}
+}
+
+func TestNoticesEnterOpensURL(t *testing.T) {
+	m := modalModel(t)
+	var opened string
+	openBrowser = func(url string) error {
+		opened = url
+		return nil
+	}
+	t.Cleanup(func() { openBrowser = defaultOpenBrowser })
+
+	m.handleNoticesKey(key("down"))
+	m.handleNoticesKey(key("enter"))
+	if opened == "" {
+		t.Fatal("enter should open the selected notice's url")
+	}
+	want := m.activeNotices()[1].url
+	if opened != want {
+		t.Fatalf("opened %q, want %q", opened, want)
+	}
+}
+
+func TestNoticesEscCloses(t *testing.T) {
+	m := modalModel(t)
+	m.handleNoticesKey(key("esc"))
+	if m.mode != modeList {
+		t.Fatal("esc should close the modal")
+	}
+	if len(m.activeNotices()) == 0 {
+		t.Fatal("esc must not dismiss anything")
+	}
+}
+
+func TestDevBuildNeverGreets(t *testing.T) {
+	st := noticeStore(t)
+	m := noticeModel(st, "dev")
+	m.openStartupNotice()
+	if m.mode == modeNotices {
+		t.Fatal("a dev build must not open the startup modal")
+	}
+	if seen, _ := st.Setting(lastSeenVersionSetting); seen != "" {
+		t.Fatalf("a dev build must not advance the stored version, got %q", seen)
+	}
+}
+
+func TestStartupOpensNoticesModalOncePerVersion(t *testing.T) {
+	st := noticeStore(t)
+	m := noticeModel(st, "v0.2.0")
+	m.openStartupNotice()
+	if m.mode != modeNotices {
+		t.Fatalf("first launch should open the notices modal, mode=%v", m.mode)
+	}
+	if selected := m.activeNotices()[m.noticeCursor]; selected.id != noticeWelcome {
+		t.Fatalf("welcome should be selected, got %q", selected.id)
+	}
+
+	again := noticeModel(st, "v0.2.0")
+	again.openStartupNotice()
+	if again.mode == modeNotices {
+		t.Fatal("a later launch of the same version must not reopen the modal")
+	}
+}
+
+func TestLateReleaseKeepsModalSelection(t *testing.T) {
+	m := modalModel(t)
+	m.handleNoticesKey(key("down"))
+	selected := m.activeNotices()[m.noticeCursor].id
+
+	m.Update(updateMsg{latest: "v9.9.9", url: "https://example.com"})
+	if got := m.activeNotices()[m.noticeCursor].id; got != selected {
+		t.Fatalf("selection moved from %q to %q when the release arrived", selected, got)
+	}
+}
+
+func TestLateDismissedReleaseKeepsModalSelection(t *testing.T) {
+	m := modalModel(t)
+	m.dismissNotice("update-v9.9.9")
+	selected := m.activeNotices()[m.noticeCursor].id
+
+	m.Update(updateMsg{latest: "v9.9.9", url: "https://example.com"})
+	if got := m.activeNotices()[m.noticeCursor].id; got != selected {
+		t.Fatalf("a dismissed release must not move the selection, went from %q to %q", selected, got)
+	}
+}
+
+func TestLateFeedKeepsModalSelection(t *testing.T) {
+	m := modalModel(t)
+	m.handleNoticesKey(key("down"))
+	selected := m.activeNotices()[m.noticeCursor].id
+
+	m.Update(feedMsg{messages: []feed.Message{{ID: "feed-late", Banner: "late", Title: "Late"}}})
+	if !contains(noticeIDs(m.activeNotices()), "feed-late") {
+		t.Fatal("feed message should have landed")
+	}
+	if got := m.activeNotices()[m.noticeCursor].id; got != selected {
+		t.Fatalf("selection moved from %q to %q when the feed arrived", selected, got)
+	}
 }
